requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
cassandra-driver==3.29.1
    # via betting_app (setup.py)
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   geomet
    #   typer
    #   uvicorn
dnspython==2.6.1
    # via email-validator
email-validator==2.1.1
    # via fastapi
exceptiongroup==1.2.1
    # via anyio
fastapi==0.111.0
    # via
    #   betting_app (setup.py)
    #   fastapi-cli
fastapi-cli==0.0.3
    # via fastapi
geomet==0.2.1.post1
    # via cassandra-driver
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via fastapi
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
numpy==1.26.4
    # via
    #   pandas
    #   pyarrow
orjson==3.10.3
    # via fastapi
pandas==2.2.2
    # via betting_app (setup.py)
py4j==0.10.9.7
    # via pyspark
pyarrow==16.1.0
    # via betting_app (setup.py)
pydantic==2.7.1
    # via fastapi
pydantic-core==2.18.2
    # via pydantic
pygments==2.18.0
    # via rich
pyspark==3.5.1
    # via betting_app (setup.py)
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via fastapi
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via uvicorn
rich==13.7.1
    # via typer
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   geomet
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.11.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typer
    #   uvicorn
tzdata==2024.1
    # via pandas
ujson==5.10.0
    # via fastapi
uvicorn[standard]==0.29.0
    # via
    #   betting_app (setup.py)
    #   fastapi
    #   fastapi-cli
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
